code/requirements.txt
mne>=1.6
numpy>=1.24
scipy>=1.11
pandas>=2.0
powerlaw>=1.5
antropy>=0.1.6
neurokit2>=0.2
scikit-learn>=1.3
matplotlib>=3.7
seaborn>=0.12
h5py>=3.9
pyyaml>=6.0
statsmodels>=0.14
